Refactor Keyboard to own component
